requirements.txt
appdirs==1.4.3
attrs==21.2.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
geographiclib==1.52
geopy==2.2.0
html5lib==1.0.1
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
nominatim==0.1
numpy==1.21.1
packaging==20.3
pep517==0.8.2
pluggy==0.13.1
progress==1.5
py==1.10.0
pyparsing==2.4.6
pytest==6.2.4
pytest-cov==2.12.1
python-dateutil==2.8.2
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
timezonefinder==5.2.0
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1